refactor(infra): split pg_store.py (1384 lines) into concern-scoped mixins
pg_store.py exceeded the local 300-line §4.1 cap by 4.6x. Split behind
the existing pg_store_host.py PgStoreHost contract pattern (the same
shape workflow_graph_source_ast.py used for #275 and
core/context_assembly/condensers.py used for #228: thin facade +
re-export, leaf modules by responsibility):

- pg_store_schema.py   — connection creation + Phase 5 pool lifecycle
- pg_store_ddl.py      — pooled query execution (_execute) + DDL
                          migration (_init_schema); module-level
                          compute_ddl_hash/read_schema_hash/
                          _get_database_url re-exported from pg_store.py
                          for mcp_server.migrate
- pg_store_serialize.py — embedding<->bytes, datetime normalization,
                          row shaping
- pg_store_write.py    — memory INSERT path (SQL constant + param
                          building + commit)
- pg_store_supersede.py — atomic reconsolidation-supersession
                          (chain-head CAS + anchor transfer inside one
                          transaction — boundary untouched)
- pg_store_heat.py     — A3 heat_base writers + homeostatic factor
- pg_store_memory_meta.py — single-row metadata writers + mood +
                          compression
- pg_store_search.py   — recall/FTS/vector search + server-side
                          signals

PgMemoryStore's public API is unchanged: same class, same method
names/signatures, same import path. pg_store_host.py's PgStoreHost
TYPE_CHECKING contract gained the new cross-mixin members
(interactive_pool, acquire_interactive, _bytes_to_vector,
_vector_to_bytes, _isoformat_datetime_fields, _insert_memory_on).

Four methods exceeded the local 40-line cap pre-split (_init_schema,
_build_insert_params, supersede_atomic, recall_memories) — each split
into named helpers with no logic change (e.g. _build_insert_params's
33-key dict now assembled from two merged sub-dicts; supersede_atomic's
per-attempt transaction body extracted to _supersede_attempt, same
transaction boundary).

_deallocate_all/_reconnect/_execute_on_conn/_init_schema keep the
explicit "mcp_server.infrastructure.pg_store" logger name (not
__name__) so log output is unchanged for any external log-name filter.

test_I2_canonical_writer.py's line-pinned heat_base writer allow-list
updated for the 3 relocated sites (bump_heat_raw, update_memories_heat_batch,
_transfer_anchor_on) — same writers, no new ones. test_s110_sweep_
infrastructure.py's register_vector monkeypatch retargeted to
pg_store_schema (where _reconnect now lives).

Verified: full suite 7233 passed / 147 skipped / 0 new failures
(2 pre-existing agent_briefing failures, unrelated — issue #400,
open fix in PR #403), ruff check + format clean, pyright zero-diagnostic
on mcp_server/, all 9 resulting files <=300 lines, all methods <=40
lines. benchmarks/reproduce.sh --quick (LongMemEval-S, LoCoMo,
BEAM-100K) byte-identical before/after: LoCoMo MRR 0.8195/R@10 0.9543,
BEAM-100K MRR 0.6558/R@10 0.8500, LongMemEval-S MRR 0.850/R@10 100%.

Filed (not fixed, out of blast radius): issue #406 — infrastructure/
importing core/ at 6 sites (pg_store_write.py's relocated
core.temporal_normalize import among them), pre-existing since
2026-07-29, also present in untouched sqlite_store.py/pg_store_near_dup.py/
pg_store_memory_reheat.py.
